Move responsibility for text splitting configuration to the index
Previously, the splitting behaviour was defined through settings on an embedding backend. If you want to split differently for different indexes, you'd need to create multiple embedding backends.

As the only thing that needs to be aware of how content is split is the VectorIndexable object itself, the logic has now moved there and any customisations can be made on the indexed type directly.